requirements.txt
requests==2.24.0
baidu_aip==2.2.18.0
